# Search for first duplicate in vectorized `unique`